Correctly skip upgrade tests

+++ internal/suite/setup.go
@@ -2,6 +2,8 @@ package suite
 
 import (
 	"context"
+	"os"
+	"strings"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
@@ -26,8 +28,13 @@ type ClusterBuilder interface {
 // Setup handles the creation of the BeforeSuite and AfterSuite handlers. This covers the creations and cleanup of the test cluster.
 // `clusterReadyFns` can be provided if the cluster requires custom checks for cluster-ready status. If not provided the cluster will
 // be checked for at least a single control plane node being marked as ready.
-func Setup(kubeContext string, clusterBuilder ClusterBuilder, clusterReadyFns ...func(client *client.Client)) {
+func Setup(isUpgrade bool, kubeContext string, clusterBuilder ClusterBuilder, clusterReadyFns ...func(client *client.Client)) {
 	BeforeSuite(func() {
+		if isUpgrade && strings.TrimSpace(os.Getenv("E2E_OVERRIDE_VERSIONS")) == "" {
+			Skip("E2E_OVERRIDE_VERSIONS env var not set, skipping upgrade test")
+			return
+		}
+
 		logger.LogWriter = GinkgoWriter
 
 		state.SetContext(context.Background())
@@ -41,6 +48,10 @@ func Setup(kubeContext string, clusterBuilder ClusterBuilder, clusterReadyFns ..
 	})
 
 	AfterSuite(func() {
+		if isUpgrade && strings.TrimSpace(os.Getenv("E2E_OVERRIDE_VERSIONS")) == "" {
+			return
+		}
+
 		Expect(state.GetFramework().DeleteCluster(state.GetContext(), state.GetCluster())).To(Succeed())
 	})
 }
